+++ cli/internal/hasura/commonmetadata/commonmetadata_test.go
@@ -66,7 +66,8 @@ func TestClient_ExportMetadata(t *testing.T) {
             "retries": 1,
             "idle_timeout": 180,
             "max_connections": 50
-          }
+          },
+	  "use_prepared_statements": true
         }
       }
     }