+++ server/src-lib/Hasura/Backends/Postgres/Connection.hs
@@ -34,6 +34,7 @@ module Hasura.Backends.Postgres.Connection
   , pccReadReplicas
   , psciDatabaseUrl
   , psciPoolSettings
+  , psciUsePreparedStatements
   , module ET
   ) where
 
@@ -354,8 +355,9 @@ getDefaultPGPoolSettingIfNotExists connSettings defaultPgPoolSettings =
 
 data PostgresSourceConnInfo
   = PostgresSourceConnInfo
-  { _psciDatabaseUrl  :: !UrlConf
-  , _psciPoolSettings :: !(Maybe PostgresPoolSettings)
+  { _psciDatabaseUrl           :: !UrlConf
+  , _psciPoolSettings          :: !(Maybe PostgresPoolSettings)
+  , _psciUsePreparedStatements :: !Bool
   } deriving (Show, Eq, Generic)
 instance Cacheable PostgresSourceConnInfo
 instance Hashable PostgresSourceConnInfo
@@ -369,6 +371,7 @@ instance FromJSON PostgresSourceConnInfo where
     PostgresSourceConnInfo
       <$> o .: "database_url"
       <*> o .:? "pool_settings"
+      <*> o .:? "use_prepared_statements" .!= False -- By default preparing statements is OFF for postgres source
 
 instance Arbitrary PostgresSourceConnInfo where
   arbitrary = genericArbitrary